Switch to OMNI node code structure with dedicated runtime
