--- node/Cargo.toml
@@ -1,54 +1,0 @@
-[package]
-name = "robonomics-node"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-clap = { workspace = true }
-color-print = { workspace = true }
-log = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-
-robonomics-primitives = { path = "../primitives" }
-robonomics-service = { path = "./service" }
-generic-runtime = { path = "./generic-runtime" }
-dev-runtime = { path = "../runtime/dev" }
-
-# substrate
-sc-cli = { workspace = true }
-sc-client-api = { workspace = true }
-sc-chain-spec = { workspace = true }
-sc-service = { workspace = true }
-sc-storage-monitor = { workspace = true }
-sc-sysinfo = { workspace = true }
-sc-telemetry = { workspace = true }
-sp-core = { workspace = true }
-sc-network = { workspace = true }
-sp-runtime = { workspace = true }
-frame-benchmarking-cli = { workspace = true }
-sp-genesis-builder = { workspace = true }
-
-# polkadot
-polkadot-cli = { workspace = true }
-
-# cumulus
-cumulus-client-cli = { workspace = true }
-cumulus-client-service = { workspace = true }
-cumulus-primitives-core = { workspace = true }
-
-[build-dependencies]
-substrate-build-script-utils = { workspace = true }
-
-[features]
-runtime-benchmarks = [
-  "sc-service/runtime-benchmarks",
-  "dev-runtime/runtime-benchmarks",
-  "robonomics-service/runtime-benchmarks",
-]
