--- node/generic-runtime/Cargo.toml
@@ -1,32 +1,0 @@
-[package]
-name = "generic-runtime"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-robonomics-primitives = { path = "../../primitives" }
-frame-support = { workspace = true }
-frame-system-rpc-runtime-api = { workspace = true }
-pallet-transaction-payment = { workspace = true }
-pallet-transaction-payment-rpc-runtime-api = { workspace = true }
-sp-api = { workspace = true }
-sp-block-builder = { workspace = true }
-sp-core = { workspace = true }
-sp-std = { workspace = true }
-sp-session = { workspace = true }
-sp-runtime = { workspace = true }
-sp-transaction-pool = { workspace = true }
-sp-inherents = { workspace = true }
-sp-offchain = { workspace = true }
-sp-version = { workspace = true }
-cumulus-primitives-core = { workspace = true }
-
-[features]
-try-runtime = []
-runtime-benchmarks = []
