--- node/rpc/core/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "robonomics-rpc-core"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-jsonrpsee = { workspace = true }
-pallet-transaction-payment-rpc = { workspace = true }
-substrate-frame-rpc-system = { workspace = true }
-sp-api = { workspace = true }
-sp-blockchain = { workspace = true }
-sp-block-builder = { workspace = true }
-sc-client-api = { workspace = true }
-sc-rpc-api = { workspace = true }
-sc-transaction-pool-api = { workspace = true }
-
-robonomics-primitives = { path = "../../../primitives" }
