--- node/rpc/experimental/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "robonomics-rpc-experimental"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-jsonrpsee = { workspace = true }
