--- node/service/Cargo.toml
@@ -1,71 +1,0 @@
-[package]
-name = "robonomics-service"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[package.metadata.docs.rs]
-targets = ["x86_64-unknown-linux-gnu"]
-
-[dependencies]
-log = { workspace = true }
-futures = { workspace = true }
-jsonrpsee = { workspace = true }
-parity-scale-codec = { workspace = true }
-
-sc-basic-authorship = { workspace = true }
-sc-client-api = { workspace = true }
-sc-consensus-aura = { workspace = true }
-sc-consensus = { workspace = true }
-sc-consensus-grandpa = { workspace = true }
-sc-executor = { workspace = true }
-sc-service = { workspace = true }
-sc-sysinfo = { workspace = true }
-sc-tracing = { workspace = true }
-sc-offchain = { workspace = true }
-sc-rpc = { workspace = true }
-sc-transaction-pool = { workspace = true }
-sc-transaction-pool-api = { workspace = true }
-sc-telemetry = { workspace = true }
-sc-network = { workspace = true }
-sc-network-sync = { workspace = true }
-sp-api = { workspace = true }
-sp-block-builder = { workspace = true }
-sp-consensus = { workspace = true }
-sp-consensus-aura = { workspace = true }
-sp-consensus-grandpa = { workspace = true }
-sp-io = { workspace = true }
-sp-offchain = { workspace = true }
-sp-runtime = { workspace = true }
-sp-session = { workspace = true }
-sp-transaction-pool = { workspace = true }
-sp-timestamp = { workspace = true }
-sp-keystore = { workspace = true }
-
-substrate-prometheus-endpoint = { workspace = true }
-frame-benchmarking = { workspace = true }
-frame-benchmarking-cli = { workspace = true }
-frame-system-rpc-runtime-api = { workspace = true }
-pallet-transaction-payment-rpc-runtime-api = { workspace = true }
-
-cumulus-client-cli = { workspace = true }
-cumulus-client-consensus-common = { workspace = true }
-cumulus-client-consensus-relay-chain = { workspace = true }
-cumulus-client-network = { workspace = true }
-cumulus-client-service = { workspace = true }
-cumulus-client-parachain-inherent = { workspace = true }
-cumulus-primitives-core = { workspace = true }
-cumulus-primitives-parachain-inherent = { workspace = true }
-cumulus-relay-chain-interface = { workspace = true }
-
-robonomics-primitives = { path = "../../primitives" }
-robonomics-rpc-core = { path = "../rpc/core" }
-dev-runtime = { path = "../../runtime/dev" }
-pallet-robonomics-lighthouse = { path = "../../frame/lighthouse" }
-
-[features]
-runtime-benchmarks = [
-  "dev-runtime/runtime-benchmarks"
-]
