--- runtime/dev/Cargo.toml
@@ -1,166 +1,0 @@
-[package]
-name = "dev-runtime"
-build = "build.rs"
-version.workspace = true
-authors.workspace = true
-edition.workspace = true
-homepage.workspace = true
-repository.workspace = true
-
-[dependencies]
-# third-party dependencies
-parity-scale-codec = { workspace = true }
-scale-info = { workspace = true }
-hex-literal = { workspace = true, optional = true }
-# serde_json = { workspace = true, default-features = false, features = ["alloc"] }
-serde = { workspace = true, default-features = false }
-serde_json = { workspace = true, default-features = false }
-
-# primitives
-robonomics-primitives = { path = "../../primitives", default-features = false }
-sp-api = { workspace = true }
-sp-block-builder = { workspace = true }
-sp-core = { workspace = true }
-sp-consensus-aura = { workspace = true }
-sp-inherents = { workspace = true }
-sp-offchain = { workspace = true }
-sp-runtime = { workspace = true }
-sp-session = { workspace = true }
-sp-std = { workspace = true }
-sp-transaction-pool = { workspace = true }
-sp-version = { workspace = true }
-sp-genesis-builder = { workspace = true }
-sp-keyring = { workspace = true }
-
-# frame dependencies
-pallet-assets = { workspace = true }
-pallet-aura = { workspace = true }
-pallet-balances = { workspace = true }
-pallet-collective = { workspace = true }
-pallet-democracy = { workspace = true }
-frame-executive = { workspace = true }
-pallet-grandpa = { workspace = true }
-pallet-identity = { workspace = true }
-pallet-preimage = { workspace = true }
-pallet-scheduler = { workspace = true }
-pallet-sudo = { workspace = true }
-frame-support = { workspace = true }
-frame-benchmarking = { workspace = true, optional = true }
-frame-system = { workspace = true }
-frame-system-rpc-runtime-api = { workspace = true }
-frame-system-benchmarking = { workspace = true, optional = true }
-pallet-membership = { workspace = true }
-pallet-multisig = { workspace = true }
-pallet-migrations = { workspace = true }
-pallet-utility = { workspace = true }
-pallet-vesting = { workspace = true }
-pallet-timestamp = { workspace = true }
-pallet-transaction-payment = { workspace = true }
-pallet-transaction-payment-rpc-runtime-api = { workspace = true }
-pallet-treasury = { workspace = true }
-
-cumulus-pallet-parachain-system = { workspace = true }
-cumulus-pallet-session-benchmarking = { workspace = true }
-cumulus-pallet-xcmp-queue = { workspace = true }
-cumulus-primitives-core = { workspace = true }
-cumulus-primitives-utility = { workspace = true }
-
-pallet-collator-selection = { workspace = true }
-pallet-message-queue = { workspace = true }
-pallet-xcm = { workspace = true }
-parachains-common = { workspace = true }
-polkadot-parachain-primitives = { workspace = true }
-polkadot-runtime-common = { workspace = true }
-
-# robonomics pallets
-pallet-robonomics-rws = { path = "../../frame/rws", default-features = false }
-pallet-robonomics-launch = { path = "../../frame/launch", default-features = false }
-pallet-robonomics-datalog = { path = "../../frame/datalog", default-features = false }
-pallet-robonomics-liability = { path = "../../frame/liability", default-features = false }
-pallet-robonomics-digital-twin = { path = "../../frame/digital-twin", default-features = false }
-
-[build-dependencies]
-substrate-wasm-builder = { workspace = true }
-
-[features]
-default = ["std"]
-std = [
-  "parity-scale-codec/std",
-  "pallet-assets/std",
-  "pallet-aura/std",
-  "pallet-balances/std",
-  "pallet-collective/std",
-  "pallet-democracy/std",
-  "sp-api/std",
-  "frame-executive/std",
-  "pallet-grandpa/std",
-  "robonomics-primitives/std",
-  "sp-offchain/std",
-  "sp-core/std",
-  "sp-consensus-aura/std",
-  "sp-std/std",
-  "sp-runtime/std",
-  "pallet-identity/std",
-  "pallet-preimage/std",
-  "pallet-scheduler/std",
-  "pallet-sudo/std",
-  "frame-support/std",
-  "frame-benchmarking/std",
-  "frame-system-benchmarking/std",
-  "frame-system/std",
-  "frame-system-rpc-runtime-api/std",
-  "pallet-membership/std",
-  "pallet-multisig/std",
-  "pallet-utility/std",
-  "pallet-vesting/std",
-  "pallet-timestamp/std",
-  "pallet-transaction-payment/std",
-  "pallet-transaction-payment-rpc-runtime-api/std",
-  "pallet-treasury/std",
-  "sp-version/std",
-  "sp-block-builder/std",
-  "sp-transaction-pool/std",
-  "sp-inherents/std",
-  "pallet-robonomics-rws/std",
-  "pallet-robonomics-launch/std",
-  "pallet-robonomics-datalog/std",
-  "pallet-robonomics-liability/std",
-  "pallet-robonomics-digital-twin/std",
-  # for benchmarks:
-  "cumulus-pallet-parachain-system/std",
-  "cumulus-pallet-session-benchmarking/std",
-  "cumulus-pallet-xcmp-queue/std",
-  "cumulus-primitives-core/std",
-  "cumulus-primitives-utility/std",
-  "pallet-collator-selection/std",
-  "pallet-message-queue/std",
-  "pallet-xcm/std",
-  "parachains-common/std",
-  "polkadot-parachain-primitives/std",
-  "polkadot-runtime-common/std",
-]
-runtime-benchmarks = [
-  "cumulus-pallet-parachain-system/runtime-benchmarks",
-  "cumulus-pallet-session-benchmarking/runtime-benchmarks",
-  "cumulus-pallet-xcmp-queue/runtime-benchmarks",
-  "cumulus-primitives-core/runtime-benchmarks",
-  "cumulus-primitives-utility/runtime-benchmarks",
-  "frame-benchmarking/runtime-benchmarks",
-  "frame-support/runtime-benchmarks",
-  "frame-system-benchmarking/runtime-benchmarks",
-  "frame-system/runtime-benchmarks",
-  "hex-literal",
-  "pallet-balances/runtime-benchmarks",
-  "pallet-collator-selection/runtime-benchmarks",
-  "pallet-collective/runtime-benchmarks",
-  "pallet-robonomics-launch/runtime-benchmarks",
-  "pallet-robonomics-datalog/runtime-benchmarks",
-  "pallet-message-queue/runtime-benchmarks",
-  "pallet-sudo/runtime-benchmarks",
-  "pallet-timestamp/runtime-benchmarks",
-  "pallet-xcm/runtime-benchmarks",
-  "parachains-common/runtime-benchmarks",
-  "polkadot-parachain-primitives/runtime-benchmarks",
-  "polkadot-runtime-common/runtime-benchmarks",
-  "sp-runtime/runtime-benchmarks",
-]
