--- runtime/dev/src/constants.rs
@@ -1,82 +1,0 @@
-///////////////////////////////////////////////////////////////////////////////
-//
-//  Copyright 2018-2024 Robonomics Network <research@robonomics.network>
-//
-//  Licensed under the Apache License, Version 2.0 (the "License");
-//  you may not use this file except in compliance with the License.
-//  You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-//  Unless required by applicable law or agreed to in writing, software
-//  distributed under the License is distributed on an "AS IS" BASIS,
-//  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-//  See the License for the specific language governing permissions and
-//  limitations under the License.
-//
-///////////////////////////////////////////////////////////////////////////////
-//! A set of constant values used in substrate runtime.
-
-/// Money matters.
-pub mod currency {
-    use robonomics_primitives::Balance;
-
-    pub const COASE: Balance = 1_000;
-    pub const GLUSHKOV: Balance = 1_000 * COASE;
-    pub const XRT: Balance = 1_000 * GLUSHKOV;
-
-    pub const fn deposit(items: u32, bytes: u32) -> Balance {
-        items as Balance * 150 * GLUSHKOV + (bytes as Balance) * 60 * GLUSHKOV
-    }
-}
-
-/// Time.
-pub mod time {
-    use robonomics_primitives::{BlockNumber, Moment};
-
-    /// Since BABE is probabilistic this is the average expected block time that
-    /// we are targetting. Blocks will be produced at a minimum duration defined
-    /// by `SLOT_DURATION`, but some slots will not be allocated to any
-    /// authority and hence no block will be produced. We expect to have this
-    /// block time on average following the defined slot duration and the value
-    /// of `c` configured for BABE.
-    /// This value is only used indirectly to define the unit constants below
-    /// that are expressed in blocks. The rest of the code should use
-    /// `SLOT_DURATION` instead (like the timestamp module for calculating the
-    /// minimum period).
-    /// <https://research.web3.foundation/en/latest/polkadot/BABE/Babe/#6-practical-results>
-    pub const MILLISECS_PER_BLOCK: Moment = 4000;
-    pub const SECS_PER_BLOCK: Moment = MILLISECS_PER_BLOCK / 1000;
-
-    pub const SLOT_DURATION: Moment = MILLISECS_PER_BLOCK;
-
-    // 1 in 4 blocks (on average, not counting collisions) will be primary BABE blocks.
-    pub const PRIMARY_PROBABILITY: (u64, u64) = (1, 4);
-
-    pub const EPOCH_DURATION_IN_BLOCKS: BlockNumber = 10 * MINUTES;
-    pub const EPOCH_DURATION_IN_SLOTS: u64 = {
-        const SLOT_FILL_RATE: f64 = MILLISECS_PER_BLOCK as f64 / SLOT_DURATION as f64;
-
-        (EPOCH_DURATION_IN_BLOCKS as f64 * SLOT_FILL_RATE) as u64
-    };
-
-    pub const MINUTES: BlockNumber = 60 / (SECS_PER_BLOCK as BlockNumber);
-    pub const HOURS: BlockNumber = MINUTES * 60;
-    pub const DAYS: BlockNumber = HOURS * 24;
-}
-
-// CRITICAL NOTE: The system module maintains two constants: a _maximum_ block weight and a
-// _ratio_ of it yielding the portion which is accessible to normal transactions (reserving the rest
-// for operational ones). `TARGET_BLOCK_FULLNESS` is entirely independent and the system module is
-// not aware of if, nor should it care about it. This constant simply denotes on which ratio of the
-// _maximum_ block weight we tweak the fees. It does NOT care about the type of the dispatch.
-//
-// For the system to be configured in a sane way, `TARGET_BLOCK_FULLNESS` should always be less than
-// the ratio that `system` module uses to find normal transaction quota.
-/// Fee-related.
-pub mod fee {
-    pub use sp_runtime::Perbill;
-
-    /// The block saturation level. Fees will be updates based on this value.
-    pub const TARGET_BLOCK_FULLNESS: Perbill = Perbill::from_percent(25);
-}
