+++ src/main.rs
@@ -1,6 +1,6 @@
 ///////////////////////////////////////////////////////////////////////////////
 //
-//  Copyright 2018-2024 Robonomics Network <research@robonomics.network>
+//  Copyright 2018-2025 Robonomics Network <research@robonomics.network>
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -15,14 +15,37 @@
 //  limitations under the License.
 //
 ///////////////////////////////////////////////////////////////////////////////
-/// Global allocator. Changing it to another allocator will require
-/// changing `memory_stats::MemoryAllocationTracker`.
-#[cfg(any(target_os = "linux", feature = "jemalloc-allocator"))]
-#[global_allocator]
-pub static ALLOC: tikv_jemallocator::Jemalloc = tikv_jemallocator::Jemalloc;
+#![warn(unused_extern_crates)]
+
+use polkadot_omni_node_lib::{
+	chain_spec::DiskChainSpecLoader, run, runtime::DefaultRuntimeResolver, CliConfig as CliConfigT,
+	RunConfig, NODE_VERSION,
+};
+
+struct CliConfig;
+
+impl CliConfigT for CliConfig {
+	fn impl_version() -> String {
+		let commit_hash = env!("SUBSTRATE_CLI_COMMIT_HASH");
+		format!("robonomics-{commit_hash} :: polkadot omni node v{}", NODE_VERSION)
+	}
+
+	fn author() -> String {
+		env!("CARGO_PKG_AUTHORS").into()
+	}
+
+	fn support_url() -> String {
+		"https://github.com/airalab/robonomics/issues/new".into()
+	}
+
+	fn copyright_start_year() -> u16 {
+		2018
+	}
+}
 
 fn main() -> color_eyre::eyre::Result<()> {
-    color_eyre::install()?;
-    robonomics_node::run()?;
-    Ok(())
+	color_eyre::install()?;
+
+	let config = RunConfig::new(Box::new(DefaultRuntimeResolver), Box::new(DiskChainSpecLoader));
+	Ok(run::<CliConfig>(config)?)
 }
